Separate PlatformsEditingView
